+++ pytorch_lightning/trainer/trainer.py
@@ -87,7 +87,7 @@ from pytorch_lightning.utilities.argparse import (
 )
 from pytorch_lightning.utilities.cloud_io import get_filesystem
 from pytorch_lightning.utilities.distributed import distributed_available
-from pytorch_lightning.utilities.exceptions import MisconfigurationException
+from pytorch_lightning.utilities.exceptions import ExitGracefullyException, MisconfigurationException
 from pytorch_lightning.utilities.imports import _fault_tolerant_training
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.seed import reset_seed
@@ -2005,6 +2005,12 @@ class Trainer(
         active_loop = self._active_loop
         if active_loop is not None:
             return active_loop._results
+
+    def _exit_gracefully_on_signal(self) -> None:
+        if _fault_tolerant_training() and self._terminate_gracefully:
+            caller = inspect.stack()[1]
+            class_name = caller[0].f_locals["self"].__class__.__name__
+            raise ExitGracefullyException(f"Exiting gracefully on {class_name}:{caller.function}")
 
     """
     Other